Create a inventory.py for readability
