+++ app/templates/profile.html
@@ -51,7 +51,7 @@
         <div class="card-body">
             <p><strong>Trades Completed:</strong> {{ player.trade_completed }}</p>
 
-            <a href="{{ url_for('main.trading') }}" class="btn btn-success">
+            <a href="{{ url_for('trading.trading') }}" class="btn btn-success">
                 Go to Trading Page
             </a>
         </div>
