+++ app/templates/trading.html
@@ -8,7 +8,7 @@
     <a href="{{ url_for('main.profile', username=session.get('username')) }}" class="btn btn-outline-primary">
         Back to Profile
     </a>
-    <a href="{{ url_for('main.new_trade') }}" class="btn btn-primary">Create New Trade</a>
+    <a href="{{ url_for('trading.new_trade') }}" class="btn btn-primary">Create New Trade</a>
 </div>
 
 <div class="row g-5">
@@ -27,7 +27,7 @@
                                 <p class="mb-1">Status: {{ trade.status|capitalize }}</p>
                                 <p class="mb-1">They want: {{ trade.cards_requested }} cards</p>
                                 <p class="mb-2">They offer: {{ trade.cards_offered }} cards</p>
-                                <a href="{{ url_for('main.view_trade', trade_id=trade.id) }}" class="btn btn-sm btn-outline-primary">View Trade</a>
+                                <a href="{{ url_for('trading.view_trade', trade_id=trade.id) }}" class="btn btn-sm btn-outline-primary">View Trade</a>
                             </div>
                         {% endfor %}
                     {% else %}
@@ -53,7 +53,7 @@
                                 <p class="mb-1">Status: {{ trade.status|capitalize }}</p>
                                 <p class="mb-1">You want: {{ trade.cards_requested }} cards</p>
                                 <p class="mb-2">You offer: {{ trade.cards_offered }} cards</p>
-                                <a href="{{ url_for('main.view_trade', trade_id=trade.id) }}" class="btn btn-sm btn-outline-primary">View Trade</a>
+                                <a href="{{ url_for('trading.view_trade', trade_id=trade.id) }}" class="btn btn-sm btn-outline-primary">View Trade</a>
                             </div>
                         {% endfor %}
                     {% else %}
