+++ app/templates/view_trade.html
@@ -9,7 +9,7 @@
 <h1 class="text-center mb-4">Trade Details</h1>
 
 <div class="d-flex justify-content-start mb-4">
-    <a href="{{ url_for('main.trading') }}" class="btn btn-primary">Return To Trading</a>
+    <a href="{{ url_for('trading.trading') }}" class="btn btn-primary">Return To Trading</a>
 </div>
 
 <div class="card mb-4">
